Fix picsum provider.
Changes to picsum in this commit
https://github.com/nlemoine/wp-cli-fixtures/commit/24cfbafaca09a6b15dc661b1be1530c3095c25ea
meant that picsum would no longer resolve to a valid image. This fixes
it by ensuring we use the corret url and provide our own custom version
of Image::image() that will ensure curl follows redirects so that picsum
works correctly.

The problem with getting the image url without the redirect is that it
depends on the redirect to generate the hash.