Add failing test for bug with parent terms.
Only category terms can be hierarchical so you should only be able to
attach a parent to a term from the category taxonomy.

What happens in the scenario from the test is that each of the terms is created as a category then
during the perist the relevant ones are updated to be tags. However, the
parent has already been set as part of the create (due to `termId` being
called by the generator during creation).

I'm not entirely sure of the best way to fix this, but hopefully this
helps in showing the issue so a fix can be made.

How this affects things for the end user is if a parent is set on a
category and points to a tag you get lots `NOTICE: PHP message: PHP
Notice:  Trying to get property 'slug' of non-object in
/var/www/html/wp/wp-includes/taxonomy.php on line 4434` in the logs.