Fix for parent category bug.
Due to an issue where tags could be used for a parent of a category, we
need to be able to set the taxonomy at construction time. This allows a
taxonomy to be passed for a term which works around the issue.

What we may want to consider to avoid confusion is to ditch the taxonomy
field altogether in favour of passing it through the constructor.